requirements.txt
openai==1.3.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pytz==2023.3
streamlit==1.28.0
pandas==2.1.3
plotly==5.17.0
